refactor: Refactors DTOs to clarify user and login data separation
Separates user entity representation from login credentials,
improving code clarity and reinforcing that passwords are not exposed
in user data structures. Updates imports for consistency and maintainability.
