+++ src/services/user.service.ts
@@ -3,8 +3,7 @@ import { User as UserEntity } from '@prisma/client';
 import * as bcrypt from 'bcrypt';
 import { hashSync } from 'bcrypt';
 import { UserRepository } from '../database';
-import { LoginDto } from '../dtos';
-import { CreateUserDto } from '../dtos/user.dto';
+import { LoginDto, CreateUserDto } from '../dtos';
 import { User } from '../models';
 import { JwtService } from './jwt.service';
 import { Result } from '../types/result.type';
